// src/modules/ecommerce/products/infra/http/entities/interfaces/IProdutoAnyShop.ts
import IDepartamentoAnyShop from './IDepartamentoAnyShop';
import IDimensoesAnyShop from './IDimensoesAnyShop';
import IEstoquesAnyShop from './IEstoquesAnyShop';
import IFreteAnyShop from './IFreteAnyShop';
import IPrecosAnyShop from './IPrecosAnyShop';
import IFabricanteAnyShop from './IFabricanteAnyShop';
import IGaleriaAnyShop from './IGaleriaAnyShop';
import IImagemPrincipalAnyShop from './IImagemPrincipalAnyShop';
import IImagemSecundariaAnyShop from './IImagemSecundariaAnyShop';
// import ILancamentoAnyShop from './ILancamentoAnyShop';
// import IPromocaoAnyShop from './IPromocaoAnyShop';
import ISeoAnyShop from './ISeoAnyShop';
// import ISugestaoAnyShop from './ISugestaoAnyShop';
// import ITextoAnyShop from './ITextoAnyShop';
// import IVideoAnyShop from './IVideoAnyShop';

export default interface IProdutoAnyShop {
  id: number;
  id_externo_24: string;
  titulo_externo_8: string;
  titulo: string;
  formato_produto: string;
  tipo_produto: string;
  ativo: number;
  cod_referencia: string;
  destaque: number;
  frete: IFreteAnyShop;
  seo: ISeoAnyShop;
  precos: IPrecosAnyShop;
  estoques: IEstoquesAnyShop;
  dimensoes: IDimensoesAnyShop;
  // promocao: IPromocaoAnyShop;
  // lancamento: ILancamentoAnyShop;
  fabricante: IFabricanteAnyShop;
  // sugestoes: ISugestaoAnyShop[];
  // videos: IVideoAnyShop[];
  // textos: ITextoAnyShop[];
  galeria?: IGaleriaAnyShop[];
  imagem_principal?: IImagemPrincipalAnyShop;
  imagem_secundaria?: IImagemSecundariaAnyShop;
  departamentos: IDepartamentoAnyShop[];
}